created pantry page - added search bar& background
