+++ mobileCapstoneProject/src/app/pantry/pantry-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { PantryPage } from './pantry.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: PantryPage,
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class PantryPageRoutingModule {}
