+++ mobileCapstoneProject/src/app/pantry/pantry.page.html
@@ -1,0 +1,9 @@
+  <ion-header collapse="condense">
+    <ion-toolbar>
+      <ion-title>My Pantry</ion-title>
+    <ion-searchbar placeholder="Add ingredients"></ion-searchbar>
+  </ion-toolbar>
+  </ion-header>
+<ion-content [fullscreen]="true">
+
+</ion-content>
