+++ mobileCapstoneProject/src/app/pantry/pantry.page.scss
@@ -1,0 +1,19 @@
+ion-header, ion-title{
+  background: url(../../assets/images/dark-wood.jpg) 0/100% 100%;
+}
+
+ion-toolbar, ion-header {
+  --min-height: 150px;
+ }
+
+ion-title{
+  color: white;
+}
+
+ion-searchbar{
+  --background: white;
+}
+
+ion-content{
+  --background: url(../../assets/images/light-wood.jpg) 0/100% 100%;
+}
