+++ mobileCapstoneProject/src/app/pantry/pantry.page.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-pantry',
+  templateUrl: 'pantry.page.html',
+  styleUrls: ['pantry.page.scss']
+})
+export class PantryPage {
+
+  constructor() {}
+
+}
