+++ mobileCapstoneProject/src/app/tabs/tabs.page.html
@@ -3,17 +3,17 @@
   <ion-tab-bar slot="bottom">
     <ion-tab-button tab="tab1">
       <ion-icon name="search"></ion-icon>
-      <ion-label>Tab 1</ion-label>
+      <ion-label>Dashboard</ion-label>
     </ion-tab-button>
 
     <ion-tab-button tab="tab2">
       <ion-icon name="heart"></ion-icon>
-      <ion-label>Tab 2</ion-label>
+      <ion-label>Favorites</ion-label>
     </ion-tab-button>
 
     <ion-tab-button tab="tab3">
       <ion-icon name="fast-food-outline"></ion-icon>
-      <ion-label>Tab 3</ion-label>
+      <ion-label>Pantry</ion-label>
     </ion-tab-button>
   </ion-tab-bar>
