+++ mobileCapstoneProject/src/global.scss
@@ -24,3 +24,4 @@
 @import "~@ionic/angular/css/text-alignment.css";
 @import "~@ionic/angular/css/text-transformation.css";
 @import "~@ionic/angular/css/flex-utils.css";
+
