+++ mobileCapstoneProject/src/index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="utf-8" />
-  <title>Ionic App</title>
+  <title>Bits to Bites Recipe Search Application</title>
 
   <base href="/" />
